+++ src/nodebox/handle/AbstractHandle.java
@@ -53,12 +53,12 @@ public abstract class AbstractHandle implements Handle {
     protected void drawDot(GraphicsContext ctx, double x, double y) {
         BezierPath p = new BezierPath();
         p.setFillColor(HANDLE_COLOR);
-        p.rect(x, y, HANDLE_SIZE, HANDLE_SIZE);
+        p.rect((float) x, (float) y, HANDLE_SIZE, HANDLE_SIZE);
         ctx.draw(p);
     }
 
     protected void drawDot(BezierPath p, double x, double y) {
-        p.rect(x, y, HANDLE_SIZE, HANDLE_SIZE);
+        p.rect((float) x, (float) y, HANDLE_SIZE, HANDLE_SIZE);
     }
 
     /**
