+++ src/nodebox/handle/FourPointHandle.java
@@ -14,8 +14,8 @@ public class FourPointHandle extends AbstractHandle {
 
     private String xName, yName, widthName, heightName;
     private DragState dragState = DragState.NONE;
-    private double px, py;
-    private double ocx, ocy, owidth, oheight;
+    private float px, py;
+    private float ocx, ocy, owidth, oheight;
 
     public FourPointHandle(Node node) {
         this(node, "x", "y", "width", "height");
@@ -30,17 +30,17 @@ public class FourPointHandle extends AbstractHandle {
     }
 
     public void draw(GraphicsContext ctx) {
-        double cx = node.asFloat(xName);
-        double cy = node.asFloat(yName);
-        double width = node.asFloat(widthName);
-        double height = node.asFloat(heightName);
-        double left = cx - width / 2;
-        double right = cx + width / 2;
-        double top = cy - height / 2;
-        double bottom = cy + height / 2;
+        float cx = node.asFloat(xName);
+        float cy = node.asFloat(yName);
+        float width = node.asFloat(widthName);
+        float height = node.asFloat(heightName);
+        float left = cx - width / 2;
+        float right = cx + width / 2;
+        float top = cy - height / 2;
+        float bottom = cy + height / 2;
         BezierPath cornerPath = new BezierPath();
         cornerPath.setFillColor(HANDLE_COLOR);
-        cornerPath.setStrokeWidth(0.0);
+        cornerPath.setStrokeWidth(0);
         drawDot(cornerPath, left, top);
         drawDot(cornerPath, right, top);
         drawDot(cornerPath, right, bottom);
@@ -64,10 +64,10 @@ public class FourPointHandle extends AbstractHandle {
         owidth = node.asFloat(widthName);
         oheight = node.asFloat(heightName);
 
-        double left = ocx - owidth / 2;
-        double right = ocx + owidth / 2;
-        double top = ocy - oheight / 2;
-        double bottom = ocy + oheight / 2;
+        float left = ocx - owidth / 2;
+        float right = ocx + owidth / 2;
+        float top = ocy - oheight / 2;
+        float bottom = ocy + oheight / 2;
 
         Rect topLeft = createHitRectangle(left, top);
         Rect topRight = createHitRectangle(right, top);
@@ -93,11 +93,11 @@ public class FourPointHandle extends AbstractHandle {
     @Override
     public void mouseDragged(Point pt) {
         if (dragState == DragState.NONE) return;
-        double x = pt.getX();
-        double y = pt.getY();
-        double dx = x - px;
-        double dy = y - py;
-        // The delta value is multiplied by 2 to create the double effect of moving
+        float x = pt.getX();
+        float y = pt.getY();
+        float dx = x - px;
+        float dy = y - py;
+        // The delta value is multiplied by 2 to create the float effect of moving
         // the top left corner down and the bottom left corner up (in the case of
         // the top left handle).
         if (dx == 0 && dy == 0) return;
